fix: move docker types to command folder
