+++ src/lib.rs
@@ -1,6 +1,5 @@
 pub mod commands;
 pub mod config;
-pub mod docker;
 pub mod state;
 pub mod store;
 pub mod util;
